+++ source/main/zettlr.js
@@ -28,7 +28,6 @@ const ignoreDir = require('../common/util/ignore-dir')
 const ignoreFile = require('../common/util/ignore-file')
 const isDir = require('../common/util/is-dir')
 const isFile = require('../common/util/is-file')
-const FILETYPES = require('../common/data.json').filetypes
 const loadCommands = require('./commands/_autoload')
 
 /**
@@ -478,34 +477,6 @@ class Zettlr {
 
   findFile (arg) { return this._fsal.findFile(arg) }
   findDir (arg) { return this._fsal.findDir(arg) }
-
-  /**
-    * Either returns one file that matches its ID with the given term or null
-    * @param  {String} term The ID to be searched for
-    * @return {Object}      The exact match, or null.
-    */
-  findExact (term) {
-    // First try the ID
-    let file = this._fsal.findExact(term, 'id')
-    // No luck? Then try the name property
-    if (!file) {
-      file = this._fsal.findExact(term, 'name')
-      let ext = path.extname(term)
-      if (ext.length > 1) {
-        // file ending given
-        file = this._fsal.findExact(term, 'name')
-      } else {
-        // No file ending given, so let's test all allowed
-        for (let type of FILETYPES) {
-          file = this._fsal.findExact(term + type, 'name')
-          if (file) break
-        }
-      }
-    }
-
-    // If any of this has worked,
-    if (file != null) this.openFile(file.hash)
-  }
 
   /**
     * Sets the current file to the given file.